Ch: Lift state to app, install react-router
