+++ package.json
@@ -42,6 +42,7 @@
     "react-bootstrap": "^1.0.0-beta.6",
     "react-dom": "^16.8.5",
     "react-icons": "^3.5.0",
+    "react-router-dom": "^5.0.0",
     "react-scripts": "2.1.8"
   }
 }
